requirements.txt
selenium
beautifulsoup4
requests
tqdm
pyaml